requirements.txt
langchain
wikipedia
langchain-community
langchain-openai
langchain-anthropic
langchain-google-genai
python-dotenv
pydantic
duckduckgo-search